+++ NEWS.md
@@ -1,6 +1,24 @@
 # forcats (development version)
 
+* `fct_count()` no longer converts implicit NAs into explicit NAs (#151).
+
+* `fct_inseq()` behaves more robustly when factor levels aren't all numbers
+  (#221).
+
+* `as_factor()` gains a logical method that always returns a factor with
+  levels "FALSE" and "TRUE" (#185).
+
+* `fct_c()`, `fct_collapse()` and `recode()` are now explicitly
+  documented as using dynamic dots
+  https://rlang.r-lib.org/reference/dyn-dots.html (@labouz, #234).
+
+* `fct_lump()` throws error if `n` or `prop` is not numeric (@robinson_es, #169). 
+
+* `as_factor()` converts ordered factor into plain factor (@robinson_es, #216).
+
 * `first2()`, a `fct_reorder2()` helper function, sorts `.y` by the first value of `.x` (@jtr13).
+  
+* fixed bug in `fct_collapse()` so it now correctly collapses factors when `group_other = TRUE` (#172), and makes `"Other"` the last level (#202) (@gtm19, #172 & #202)
 
 # forcats 0.4.0
 
@@ -8,8 +26,6 @@
 
 * `fct_collapse()` gains a `group_other` argument to allow you to group all 
   un-named levels into `"Other"`. (#100, @AmeliaMN)
-  
-* fixed bug in `fct_collapse()` so it now correctly collapses factors when `group_other = TRUE` (#172), and makes `"Other"` the last level (#202) (@gtm19, #172 & #202)
 
 * `fct_collapse()` now accepts a `other_level` argument, to allow a user-specified `Other` level (@gtm19, #194)
 
@@ -76,7 +92,7 @@
 
 * `fct_explicit_na()` also replaces NAs encoded in levels.
 
-* `fct_lump()` correctly acccounts for `NA` values in input (#41)
+* `fct_lump()` correctly accounts for `NA` values in input (#41)
 
 * `lvls_revalue()` preserves NA levels.